+++ CHANGELOG.md
@@ -1,4 +1,36 @@
-# [0.3.4] - New Extensions
+# [0.3.5] - New Extensions
+
+- Returns the month name of the `DateTime` (in `String` format).
+
+```dart
+getMonthFromDate(locale:'en')
+```
+
+- Returns the first day of the month from the provided `DateTime` (in `String` format).
+
+```dart
+firstDayOfMonth(locale:'en')
+```
+
+- Returns the last day of the month from the provided `DateTime` (in `String` format).
+
+```dart
+latsDayOfMonth(locale:'en')
+```
+
+- Returns left side of the `String` starting from `char`.
+
+```dart
+leftOf(String char)
+```
+
+- Returns right side of the `String` starting from `char`.
+
+```dart
+rightOf(String char)
+```
+
+## [0.3.4] - New Extensions
 
 - `formatWithMask` now accepts optional parameter `specialChar` which defaults to '#'